Suggest async-websocket instead.
I was running into problems with logging due to the fact that celluloid was not handling UTF-8 byte order mask (or something equivalent coming from Slack messages), it was just plain crashing when I tried to send data to a log provider. 

I found out that celluloid is not maintained anymore and it's [up for grabs](https://github.com/celluloid/celluloid/issues/779). 

Once I checked the cousin repo https://github.com/slack-ruby/slack-ruby-client#installation I noticed that 4 adapters are supported but the one `async-websocket` is the one recommended.